Rename mobile-index to reward-index
